move image comparison test helper into internal test package
This makes it available for testing packages that the public
test package depends upon (like theme).
