+++ internal/test/testdata/diffing_master.png
