+++ internal/test/testdata/master.png
